requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
idna==2.9
Naked==0.1.31
pycrypto==2.6.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
shellescape==3.8.1
sqlparse==0.3.1
urllib3==1.25.8
